requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
PyJWT
requests
twilio==5.7.0
urllib3
psycopg2
